requirements.txt
py-cord
jishaku
pymongo[srv]
requests
discord-py-slash-command
python-dotenv
temperature-converter-py
buttons
pillow
discord-together
pyfiglet
certifi